Convert processors/document.rb to an override and add test for overrides
